change field custom_size in model statistics

+++ Carpet/models.py
@@ -84,7 +84,7 @@ class Statistics(models.Model):
     service=models.ForeignKey(Service,on_delete=models.CASCADE)
     kind=models.CharField(max_length=256,blank=False,null=False)
     size=models.CharField(max_length=256,blank=False,null=False)
-    custom_size=models.IntegerField()
+    custom_size=models.FloatField()
     
     def __str__(self) -> str:
         return self.kind
